Bank Progress, Cleanup, InventoryBuilder rework
Banks are being worked on, menus work and all and they can be tested by right-clicking on stone.  I would have finished it and committed it in one commit but there's a bunch of other stuff changed too and refactored so I instead chose to split it up.

Bits and bobs of code have been optimized all around the project.  Look there for changes, although most of the functionality should remain the same.

Inventory Builders work slightly different now.  You have to specify whether you want the inventory to close when you click on an item in the inventory.  Similarly, you can now customize it a little more by not having it cancel clicking on an item in the inventory.  Used in something like banks, where parts of it should act like a normal bank but should have customizable controls.

This project isn't dead tho lol.
If you're still reading this, that's great.
I expect nobody to read git messages cough cough craftydart cough cough, but they're useful in documenting changes to code in one centralized place.  Works better if you actually have a detailed commit message.

f

+++ game/src/main/java/com/aegeus/game/WorldManager.java
@@ -42,9 +42,11 @@ public class WorldManager {
      */
     public boolean removeWorld(File path) {
         File[] files = path.listFiles();
-        for (File file : files) {
-            if (file.isDirectory()) removeWorld(file);
-            else file.delete();
+        if(files != null) {
+            for (File file : files) {
+                if (file.isDirectory()) removeWorld(file);
+                else file.delete();
+            }
         }
         return path.delete();
     }
